+++ src/Products/PloneMeeting/config.py
@@ -194,6 +194,7 @@ MEETING_CONFIG = "meeting_config"
 
 TOOL_ID = "portal_plonemeeting"
 TOOL_FOLDER_CATEGORIES = 'categories'
+TOOL_FOLDER_CLASSIFIERS = 'classifiers'
 TOOL_FOLDER_SEARCHES = 'searches'
 TOOL_FOLDER_RECURRING_ITEMS = "recurringitems"
 TOOL_FOLDER_ITEM_TEMPLATES = "itemtemplates"
@@ -210,7 +211,7 @@ DEFAULT_COPIED_FIELDS = ['title', 'description', 'detailedDescription', 'motivat
                          'groupsInCharge', 'proposingGroupWithGroupInCharge', 'copyGroups']
 # extra fields kept when an item is cloned in the same meeting config,
 # so not the case when sent to another meeting config
-EXTRA_COPIED_FIELDS_SAME_MC = ['associatedGroups', 'category',
+EXTRA_COPIED_FIELDS_SAME_MC = ['associatedGroups', 'category', 'classifier',
                                'optionalAdvisers', 'otherMeetingConfigsClonableTo',
                                'otherMeetingConfigsClonableToPrivacy', 'oralQuestion',
                                'toDiscuss', 'privacy', 'pollType', 'textCheckList']
@@ -299,6 +300,8 @@ ITEM_INSERT_METHODS = OrderedDict((
     ('on_list_type', ['field_listTypes']),
     # according to category order;
     ('on_categories', ['category']),
+    # according to classifier order;
+    ('on_classifiers', ['classifier']),
     # according to proposing group order;
     ('on_proposing_groups', ['organization']),
     # according to all groups (among proposing group AND