+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -691,6 +691,7 @@ class MeetingConfigDescriptor(Descriptor):
 
         # MeetingConfig sub-objects --------------------------------------------
         self.categories = []  # ~[CategoryDescriptor]~
+        self.classifiers = []  # ~[CategoryDescriptor]~
         self.recurringItems = []  # ~[RecurringItemDescriptor]~
         self.itemTemplates = []  # ~[ItemTemplateDescriptor]~
         self.annexTypes = []